+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Added
-- provideMemproofs endpoint
+- `provideMemproofs` endpoint
+- `allowEmailPasswordAuth` authOpts fields
 
 ## [0.6.19-prerelease] - 2024-03-26