+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `provideMemproofs` endpoint
 - `allowEmailPasswordAuth` authOpts fields
+- `hasMemproofs` field in `AgentState`
 
 ## [0.6.19-prerelease] - 2024-03-26