+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.6.20-prerelease] - 2024-07-10
 ### Added
 - `provideMemproofs` endpoint
 - `allowEmailPasswordAuth` authOpts fields